Added new attribute to use a custom node name
That attribute will make the node look at certain elements like canId, usage, etc.  This is so we don't need to make a new case for every different type of hardware or object.  Also started to work on checkbox but couldn't get to save.  Need to add support for primitive booleans and ranges on our xsd types like canId.